Make creation of file descriptor ranges faster
During daemonization, we need to close all open file descriptors, except
for the ones that correspond to the standard streams and the ones that
the user has explicitly specified that they must be preserved. Since
these file descriptors are passed to `os.closerange()', we need to
construct a list of file descriptor ranges.

The previous way of creating these fd ranges operated on a list
of fds up to the theoretical maximum, which it then sorted and iterated.
In many environments, the highest possible fd is 1048576, which can make
this method costly and take a lot of time.

This commit speeds up the process of creating fd ranges by iterating the
list of *excluded* file descriptors instead, which is usually very
small.

Signed-off-by: Alex Pyrgiotis <apyrgio@arrikto.com>